Move CCController classes to platform
