+++ cocos/Android.mk
@@ -91,6 +91,7 @@ cocos2d.cpp \
 2d/CCAutoPolygon.cpp \
 3d/CCFrustum.cpp \
 3d/CCPlane.cpp \
+platform/CCController.cpp \
 platform/CCFileUtils.cpp \
 platform/CCGLView.cpp \
 platform/CCImage.cpp \
@@ -112,7 +113,6 @@ base/CCAsyncTaskPool.cpp \
 base/CCAutoreleasePool.cpp \
 base/CCConfiguration.cpp \
 base/CCConsole.cpp \
-base/CCController.cpp \
 base/CCData.cpp \
 base/CCDataVisitor.cpp \
 base/CCDirector.cpp \
