+++ cocos/base/CCEventListenerController.cpp
@@ -26,7 +26,7 @@
 #include "base/CCEventListenerController.h"
 #include "base/CCEventController.h"
 #include "base/ccMacros.h"
-#include "base/CCController.h"
+#include "platform/CCController.h"
 
 NS_CC_BEGIN
