+++ cocos/platform/CCController.h
