+++ cocos/platform/CMakeLists.txt
@@ -103,6 +103,7 @@ include_directories(
 
 set(COCOS_PLATFORM_SRC
 
+  platform/CCController.cpp
   platform/CCSAXParser.cpp
   platform/CCThread.cpp
   platform/CCGLView.cpp
